Allow collapsing file changes
Currently there is no way to mark files as viewed, or collapse large ones of no interest (especially generated files). This is pretty painful when you want to go back forth between several files.

This adds a chevron (same as used in the StatusChecks card) that can be clicked to toggle collapsing.