Fix backwards diff view of commits
When generating a diff view for committed changes, the previous logic compared left=commit with right=commit^. This is backwards, and led to commits appearing to do the opposite.

For other types of diff view (ex. HEAD/UNCOMMITED_CHANGES), it was correct to have "commit" on the left, because right was something with more changes than commit.

Tried to rename / refactor things a bit to make the behavior more clear too (previously several related variables were being changed with ternary operators, making it hard to understand at a glance the actual result in different scenarios).