Back out "linearize control flow in EdenServer::mount"
Summary:
We've seen crashes on Windows when the mount is initialized but a later step at
mount time fails. In between these 2 events, if a write to the working copy
happens the ProjectedFS code assumes that the InodeMap is initialized and thus
tries to access the rootInode. This assumption is unfortunately broken by the
call to mountFinished and the code should instead follow the regular unmount
code path to make sure that no pending notifications are inflight.

I suspect NFS and FUSE also would exhibit the same issue.

Reviewed By: kmancini

Differential Revision: D46169999

fbshipit-source-id: ed1eb17dc2f084acb7e2d88c555150b3b8c6662e