tooltip: stop using <Component /> to better support closures
Summary:
I noticed that the scrollbar state or `useState` state in the <Tooltip /> dropdown
does not preserve across state changes. But they preserve fine without <Tooltip />.

I think what happens is, in this function:

  function PartialSelectionAction({file}: {file: UIChangedFile}) {
    const getPanel = () => {
      return <PartialSelectionPanel file={file} />;
    };

    return (
      <Tooltip
        component={getPanel}
        ...>
      </Tooltip>
    );
  }

The `getPanel` is a dynamically created function, and gets used by Tooltip like
`<getPanel ... />`. Then React notices that `getPanel` changes every time per
render (because `getPanel` is not a "static" function) and treat the `getPanel`
as "incompatible" component and re-render the whole thing. Note the `getPanel`
needs the `file` in the closure, so it cannot be a static function.

Fix this issue by avoiding treating `component` as a React component (so React
won't treat it as "component type changed" when re-rendering).

To avoid misuse, I updated `component` signature to no longer be compatible
with a React component. Callsites need to migrate `component={Foo}` to
`component={(dismiss) => <Foo dismiss={dismiss} />}`.

Reviewed By: evangrayk

Differential Revision: D46244645

fbshipit-source-id: 40ea4c0d6637da19ac9fe2b23d9c334e2575a148