copytrace: exclude files that are not in the dest manifest
Summary:
the previous change broke `test-fb-ext-copytrace.t` tests, but somehow the failure didn't show up in the CI, this is to fix the that.

The issue happened when rebasing 8f852f229 (mv a b) to 046961699 (del a), and file `a` doesn't in the dest commit. The old logic is to keep b in the final commit, and my previous abort with an error that a is not in dest. I think a better way is to report a conflict and ask users to resolve it. But for now, I will keep the old behevior.

```
  @  046961699  72 seconds ago  zhaolong
  │  del a
  │
  │ o  8f852f229  107 seconds ago  zhaolong
  ├─╯  mv a b
  │
  o  7a737d7e6  2 minutes ago  zhaolong
╭─╯  add a
```

Reviewed By: sggutier

Differential Revision: D46357844

fbshipit-source-id: f5860c0037b909e2dd1a62700775dba505b88849