filemerge: add start of simple merge tool DSL
Summary:
A user wants to use a different merge tool for delete+modified conflicts. The current merge tools are not flexible or composable enough to handle this. For example, even if we added a ":delete-keep-local" tool, there is no way to specify the fallback merge tool to use for other kinds of conflicts.

In this diff I introduce a simple DSL to solve the above problem. I called it "merge tool script". It only supports the "if" function, and "isabsent" to test of the "local" or "other" file context is absent/deleted.

Reviewed By: zzl0

Differential Revision: D46163270

fbshipit-source-id: 28f2579e665dc9dd66ab816b10e0b7a758e772f5