debugnamecomplete: show 9 digits for autocompleted rev numbers
Summary: Show 9 characters for autocompleted revs, to match rev numbers shown in e.g. `hg sl`. This makes it a little easier to verify the rev that was autocompleted for you matches the one you were looking at in the smart log (i.e. if you're looking at the end of the rev rather than the beginning).

Reviewed By: zzl0

Differential Revision: D46398818

fbshipit-source-id: 70f19d60df7aa85371e1c1838b362b621da5b493