revset: kill _firstdescendants
Summary:
It's already the same as `descendants` in our current implementation.
It's non-trivial to implement efficiently and is rarely used. So let's just
remove it for now.

Note the order of children of a commit is undefined, unlike parents. If we want
to provide a "meaningful" order and implement this in the future we might want
to use commit date etc to sort children.

Reviewed By: zzl0

Differential Revision: D45936979

fbshipit-source-id: bff72a130a5ffb7c573f06374da8f9c02791049f