prjfs: move PrjfsChannel::start onto PrjfsChannel::initialize
Summary:
In the case where PrjfsChannel::start failed to initialized, EdenFS would crash
due to the stopPromise_ not being fulfilled. That promise is being set in the
PrjfsChannel::stop method, but unfortunately, when start throws, the caller
wouldn't catch the exception.

To fix this, we can simply move all the code in the initialize method, which if
it throws, the caller already properly tears down the mount thus solving the
crash mentioned above.

Reviewed By: genevievehelsel

Differential Revision: D46210460

fbshipit-source-id: 1524e8e0fc74ce3119b1b6dcd35b5419efcebdf3