eliminate completely unnecessary traffic to commit graph endpoint
Summary:
eliminate completely unnecessary traffic to commit graph endpoint for
main bookmark (for segmented changelog supported repos, they use pull_fast_forward_master edenapi endpoint)

After the fast path pull being completed, the filtration of what is already
present wasn't working correctly. It could easily skip (not to be able to
filter out) what was just fetched via the fast path.

That resulted in completely unnecessary traffic to commit graph (which is
slower, especially for large bits of the master branch).

During the SEV that traffic was super expensive due to the www merge commit, but even without SEV this traffic leads to unnecessary load on our services especially if the repo is significantly behind master.

The solution is to invalidate caches.

Differential Revision: D46590498

fbshipit-source-id: c8d920f4ffd3411297ca7cd01368b731f3c2236e