copytrace: disable dagcopytrace for source missing files
Summary:
The dagcopytrace was enabled for source missing file in D46315431.

It's too slow if there are many files deleted in the source commit. I will debug the root cause tomorrow, let's disable it for hg repo first. We are not disabling it for git repo, because there is not alternative for git repos.

Reviewed By: muirdm

Differential Revision: D46711460

fbshipit-source-id: 4e20bbb34be126c13af081475953b83df61872df