tests: fix test-status-mlog.t on Windows
Summary:
The blackbox was reporting a mode mismatch. The cause was the filesysytem walk yielding a mode with permission bits set (e.g. 0666), but os.lstat within dirstate.normal() yields a mode without permission bits set (e.g. 0000). Fix by only reporting a mode change when the executable bit changes (which seems like the intention of the code).

I don't think this is a recent change in behavior, so I'm not going to look further into it. Hopefully the Python filesystem code will go away soon with the Rust status.

Reviewed By: zzl0

Differential Revision: D46848722

fbshipit-source-id: cdd73137ae891bc8038d7d7ddcd09d8c95170051