non-oss work 80/n
Summary:
[plan_deletion] Add subcommand to redact relevant blobs

To make it easy to use in the context of plan deletion, we take the provisional deletion plan issued from the `propose-deletion-plan` subcommand in and we output the redaction id to a file.

In the real world, one would use this redaction id to make a diff to the mononoke config in configerator and restart the mononoke servers to make sure that the redaction actually takes effect before moving on to the next step.

Reviewed By: markbt

Differential Revision: D46768075

fbshipit-source-id: 368dafd485a02d677f3a74c6ede59a80cbe931d4