Expose fetch-key-list through the API and the CLI
Summary:
The current interface relies on the assumption that redacted content only relates to file contents.
`list`, which is the way to read what was redacted discovers the blobs for filecontents by walking the fsnodes manifests.

For a feature I'm working on, I need to be able to redact arbitrary blobs, so I need a way to identify what was redacted
without relying on some derived data blobs being available.

Reviewed By: markbt

Differential Revision: D46806525

fbshipit-source-id: 78b5470d4dd741538e3e85353c0b1634f3b83b1c