nodeipc: implement fd sending on POSIX
Summary:
Make it possible to send fds on POSIX. This is similar to what chg
(hgclient.c:attachio and osutil.c:recvfds) does. Note the `sendfd`
crate seems to have a solid implementation but its public API is
on the `std::os::unix::net::UnixStream` type, which is incompatible
with our struct that uses the `FileDescriptor`.

Example:

  % cargo run --example spawn_sendfd
  Parent: spawning child
  Parent: sending hello
  Parent: sending stdio and a.txt file descriptors
  Parent: waiting for child to exit
  Child: started with IPC handle 4
  Child: got message Some("hello")
  Child: got sendfd payload SendFdPayload { raw_fds: [3, 5, 6, 7] }
  Child: writing "something\n" to fd 3
  something
  Child: writing "something\n" to fd 5
  something
  Child: writing "something\n" to fd 6
  something
  Child: writing "something\n" to fd 7
  % cat a.txt
  something

Reviewed By: muirdm

Differential Revision: D46811163

fbshipit-source-id: 37686327f02a7fa0b0e8768d025094a7fcb0e2b5