nodeipc: avoid closing underlying fd multiple times
Summary:
The `FileDescriptor` type will close the underlying fd on drop. In our
use-cases `r` and `w` refer to the same fd so we only need to close it
once. Note the reason we use separate `Mutex`es for r and w is that we
want `recv()` and `send()` running from different threads to not block
by each other.

Reviewed By: zzl0

Differential Revision: D46885308

fbshipit-source-id: b2a5644251ac4773581a995ae8c3d955b0b36861