scmstore: add lfs "passthrough" mode
Summary: Tweak things so scmstore is compatible with the python lfs extension. In particular, if remotefilelog lfs support isn't enabled, don't treat lfs pointers specially (i.e. just store them as regular file data). This is what contentstore does, and it is required for compatibility tests that use remotefilelog with the python extension. (This is not used in production.)

Reviewed By: zzl0

Differential Revision: D46674241

fbshipit-source-id: faa5b5b992a7e2cc61bf98ba2919753e9825bb0a