scmstore: fix LFS compat w/ Python LFS extension
Summary: This fixes compatibility if both the Python LFS extension and remotefilelog.lfs are enabled. In particular, if we find an LFS pointer in the non-LFS indexedlog storage (i.e. inserted by Python LFS extension when remotefilelog.lfs was not enabled), we now upgrade it to the the scmstore pointer store.

Reviewed By: quark-zju

Differential Revision: D46714084

fbshipit-source-id: 4761f1c05017897544b33b0c322d4c3bce81274d