undo: fix compat w/ debugruntest
Summary:
Use a different strategy to avoid "undo" on nested runcomand calls.

Previously it would remove the runcommand wrapper, but that effectively disables the undo extension for subsequent commands under debugruntest. Instead, we now have some global state that we unset when a command finishes so we can tell the difference between nested commands in the same process, and serial commands in the same process.

Reviewed By: zzl0

Differential Revision: D46646009

fbshipit-source-id: 47387c412da246e13b2b7c947ca04f3e419f5a87