status: enable Rust status for test-update-unknown-files.t
Summary:
I had to fix a minor issue w/ in the non-watchman case where we were case normalizing an on-disk file name w/ a dirstate entry incorrectly. When the dirstate entry (e.g. "foo") is marked for removal, it shouldn't coallesce with a different case file (e.g. "FOO"). Otherwise, it is impossible to properly show the status:

```
$ hg st
R foo
? FOO
```

Reviewed By: quark-zju

Differential Revision: D46683382

fbshipit-source-id: 83f93209ab10098efeb65fba3ba7379107ded50d