log popularity metric errors
Summary:
We started seeing lots of these errors recently. This cripples the consistent routing.

I think this is the reason we still route many blobs 1-1 even though with config we should be routing all of them to at least 2 tasks: https://fburl.com/scuba/mononoke_lfs/unjp8uue.

EDIT: well actually I think it might not be as big of an issue as I originally though. Seems like the scuba query is wrong because of the way we generate keys: https://www.internalfb.com/code/fbsource/[4b326f5dff7b0d223bdaf91f6f5649f2bc464ed6]/fbcode/eden/mononoke/lfs_server/src/batch.rs?lines=295-305. The base case, when we route 1-1, we generate routing key of form $hash. When we enable consistent routing we generate urls of the form $hash-$n (where $n is random num from [1-N)). Notice it's not 0, because if we randomly choose 0 then we just have $hash. This is not to rehash objects that are transitioning from being routed to one host to many hosts.

Might be worth to fix the errors anyway.

Reviewed By: clara-9

Differential Revision: D47209095

fbshipit-source-id: 0af315e532ffca5f6b5df08e1d89e4e214661c71