Fix bad submit as draft tooltip pluralization
Summary: This tooltip was saying "Whether to submit these 0 diffs as a draft" if you were making a new commit. That's never correct, we should just update the number to 1 if it would have been 0. This happens when making a new commit because our list of commits to submit is empty.

Reviewed By: quark-zju

Differential Revision: D47238271

fbshipit-source-id: 2585386f8c717e34f0bbac6c67b985267b4891a3