Tweak text for last pull time
Summary:
We would say "last pulled 2 hours ago" or similar when you hover the pull button. Technically, this is not true. The time we show is the newest commit in the history's age.

Internally, commits in large repos come fast enough that you probably wouldn't notice the difference between "last pull" and "newest commit". But in OSS repos, it's quite easy to have the newest commit much older than when you last pulled.

So more precise wording here makes it seem less wrong.

Reviewed By: quark-zju

Differential Revision: D47238973

fbshipit-source-id: c5594197f35d37ee48b8683ca74f0a79704aaca0