Dropdown button component
Summary:
Add a dropdown button component. This behaves like a button with a selection dropdown chevron next to it. Changing the selection via the chevron's dropdown menu changes the button text. Then, clicking the button runs that particular action.

This lets us have buttons which have built-in options / multiple valid choices. For example, we can use this for changing the submit button to Submit Draft for a draft PR/Diff.

This component is implemented as a controlled component, so the caller tracks the current selection and update the state outside the component. Typically, we'll want to persist your button choice or otherwise use an atom for it, so the responsibility is in the caller.

Reviewed By: quark-zju

Differential Revision: D43557610

fbshipit-source-id: 557418a2cf0a42255355c65fa6a21f439ac129c3