add FilteredBackingStore skeleton
Summary:
We are adding a feature that allows users to exclude or "filter" certain paths from appearing in their checkout.

This diff starts prototyping what this may look like by introducing the concept of a FilteredBackingStore. The FilteredBackingStore (FBS for short) will wrap around existing BackingStores and handle filtering out certain trees/blobs from the checkout. It achieves this by avoiding loading any paths that are children of "filtered" paths listed in the filter file.

It's TBD how filter files will be defined and where they will live. We will simply mock them in tests for now to ensure that the filtering logic works.

Reviewed By: kmancini

Differential Revision: D46495637

fbshipit-source-id: cd48ef18b5a94a438a4c1c9ea3e5696e68c62995