Improve TreeQuery performance (#673)
Summary:
[ReviewStack] Improve TreeQuery performance

Seems like "Tree.object.oid" was unused, but the fan-out of Object -> 1k Trees -> 1k Objects (repeats of the original object) added a lot of overhead.

Running queryies in github's graphql explorer, this seems to avoid a timeout.

Probably there are better improvements to be made here, but seems like a free win to include

Pull Request resolved: https://github.com/facebook/sapling/pull/673

Reviewed By: evangrayk

Differential Revision: D47340649

Pulled By: quark-zju

fbshipit-source-id: 1227f0ea0fbff97ed713d5e430a2bc812ba4ba30