Fix github check status logic (#651)
Summary:
Fix github check status logic

The current logic for github checks uses "checkSuites" to get all the check suites run for the most recent commit. This doesn't actually match which checks are shown overall for the PR (ex. some may be duplicated and cancelled, and thus not be important, or may not be blocking).

In particular, I noticed for my repo that there is some QUEUED "checkSuites" which don't appear in the github UI (copilot4prs and dependabot), and there are some cancelled checkSuites which were superseded by a later run, and thus are acceptable to fail. This leads to passing PRs showing as "pending", and some PRs showing as "failing" if a run was cancelled and rerun for the latest commit.

This PR updates to use `commit.statusCheckRollup.state` instead of manually trying to aggregate the check suites. If the breakdown is desired, `commit.statusCheckRollup.contexts.checkRunCountsByState` could provide that (but the aggregated "state" seems to handle it well).

Pull Request resolved: https://github.com/facebook/sapling/pull/651

Test Plan:
This seems to give accurate results within my repo (tested with some pending, passing, and failing cases)
<img width="146" alt="Screen Shot 2023-06-14 at 12 34 39 AM" src="https://github.com/facebook/sapling/assets/77301670/6425f673-4390-4493-8215-deead39f2eac">

Reviewed By: quark-zju

Differential Revision: D47341477

Pulled By: evangrayk

fbshipit-source-id: 1a746c2d23b45ea26cc9c1129e71738a548ec72d