Add pull --rev operation
Summary:
To support the "download commit" feature, here we add an `sl pull --rev` operation.

This is not perfect, because it doesn't support diff numbers like D12345 or PR numbers like #123. That's an artifact of using `SL_AUTOMATION=1` env var, which disables diff number lookup.

In the future, we should add an ISL-specific revset or something that lets us bypass this. For now, I'll add it to support downloading commits by hash or revset.

Reviewed By: quark-zju

Differential Revision: D47272011

fbshipit-source-id: 005dde1991ec65475fdf06fdaca646a4e9ee140d