Add Download commits menu to top bar
Summary:
Add a new top level menu dropdown which includes various "downloading commits" related features. For starters, this will be `hg pull --rev` / `jf get`.

But this can also include:
- commit cloud sync status / sync now button
- change commit cloud workspace
- changing your active main branch...?

-----

I'm hoping this button will be intuitive enough, especially with a tooltip. This is different from ISL 1.0, where the jf get funciton was an input box at the top of the screen at all times. Personally I felt that jf get is not THAT common to require that much screen space. Instead, putting it one click away keeps it in a menu along side similar features like commit cloud, which seems appealing to me.

Another option might be to try to fit jf get inside the arc pull / sl pull dropdown button. But that didn't seem quite right to me either, I like leaving that button as a dedicated pull button.

To accomodate the cloud-download icon, i changed the pull button's icon to "repo"... not sure if that makes sense but I found it weird to have TWO cloud-download icons so close to each other and mean different things.

Reviewed By: quark-zju

Differential Revision: D47272010

fbshipit-source-id: 657b40f8549ed2830470dff888e45ae6d6db646d