Preparatory work for deriving RawGitTree
Summary:
This diff includes two changes as a preparation step for deriving `RawGitTree`s:
- Make the `upload_git_object` take a `Blobstore` argument instead of using a `impl RepoBlobstore`. This will allow the git tree derivation module to reuse this functionality
- Change the `blobstore_key` for the current Git Tree type

Note that this by itself will not make the storing git tree object reusable since currently it lives in `mononoke_api` crate and referencing it in `git` crate leads to a cycle. In the next diff I will move this logic in the `git` crate so that it can be used without cyclic dependencies.

Reviewed By: markbt

Differential Revision: D47315214

fbshipit-source-id: fbc8787633aaa38089e17f005750f8dbe4b03a15