smartset: delegate filteredset.prefetch to its "subset"
Summary:
The fitleredset iteration happens on its "subset". Prefetch settings should
apply to the set that is iterated. So let's redirect prefetch APIs to the
subset.

The old `smartset.clone` implementation does not seem to preserve ordering.
The `prefetch` is updated to change prefetch fields of the `self` set directly
instead.

Reviewed By: zzl0

Differential Revision: D47419749

fbshipit-source-id: 26f6b927db5b4e7169f0d3f7e3766d5f7b696be3