dag: preserve VerLink on open or persist
Summary:
On IdDag/IdMap open, try to reuse previously assigned VerLink.
On IdDag/IdMap persist/flush, cache the storage version with the current VerLink.

This should reduce "compatibility" issues when the Python world invalidates and
re-creates the changelog, as in `dbsh`:

  In [4]: v1=repo.changelog.dag.version()
  In [5]: repo.invalidatechangelog()
  In [6]: v2=repo.changelog.dag.version()
  In [7]: v1.cmp(v2)
  Out[7]: 0

For example, take the example in D47383634 summary, now

  r1.cmp(r2)

evaluates to `-1` instead of `None` meaning that r2 has
"append-only" changes on top of r1. Although we cannot
yet get rid of D47376881.

Reviewed By: zzl0

Differential Revision: D47488534

fbshipit-source-id: 33e9ec6294ce77f85b04d5d3af02d81a1a6c9282