Add inline blame annotations
Summary:
Add a mechanism for vscode to render blame annotations next to your cursor. This is similar to the popular gitlens extension.

The annotation says how long ago the change was plus the title.

This information is fetched from `sl blame`. Unfortunately, the data format we currently have in `sl blame` is not quite enough, so we also do an `sl log` to fetch the missing info (commit description).

This blame information is only accurate at a particular commit and takes about 1s to fetch on a large repo internally. The log call takes another 0.5-1.0s. We should probably add a mechanism for `sl blame` to include this extra info, and ideally also use a more compact representation where we normalize so we have a list of hashes and one copy of the rest of the commit's info. This would give us a nice speedup. As is, at only ~2-3s per blame fetch in a new file, you don't notice the slowness that much.

In order to perform well, we only fetch blame once per file. This fetch is done on the public ancestor of the current commit. So the blame fetch does not include any local changes. This is important as it allows our blame call to be faster.

To account for local changes, we also run `sl cat` on the file at for the public commit at the base of your stack. Then, in memory we diff your current file contents with this public verison. This difference is used to realign the lines of blame. We insert empty blame lines that say `(you) local changes` for any newly added lines. This lets us account for local changes without running blame very often at all.

I consider this implementation still experimental. I'm adding this since alex-statsig put some effort into a PR to add blame:
https://github.com/facebook/sapling/pull/640

TODO:
- we need to test the reliability of this. It's worked for us internally, but this implementation differs slightly and may not work or have unexpected performance issues. As such, it's hidden behind a setting by default. Eventually we should make this setting default to ON.
- People may want to configure what's shown in the blame line. We do not put author names inline right now, they require a hover.
- The blame hover just shows the description, and a diff number. Need to verify if this works for pull requests as well.
- It seems that extension activation is based on the ISL webview or sidebar. But this feature depends on only a file being open. We should probably activate on startup finished instead of ISL being shown.
- A toggle-able larger blame that shows next to every single line would also be nice.

Reviewed By: quark-zju

Differential Revision: D47385148

fbshipit-source-id: 71fd829b4441425ac4fb1b0f999a642502a57ded