Use diff-sequences for blame diffing
Summary:
Extract the `diffLines` util from linelog into shared so we can use it for blame too. It should be a faster version of diff, in line with the results from D43857949.

The format we get is slightly different, so we have to adjust how we do the blame realignment. One easy way is to just use `.splice` to manipulate the blame array in-place for each changed hunk. I haven't measured the perf of this change before and after, but I prefer being consistent about our diff utils anyway. It's still quite snappy in my editor even on relatively large files.

Reviewed By: quark-zju

Differential Revision: D47415365

fbshipit-source-id: b70a80c2399eaa4cb37a690c96ab2b4a8808ddae