linearize control flow in EdenServer::mount take 2.1
Summary:
My previous version of this change in D44699195 was too aggressive. It
assumed mountFuture could be called unconditionally after the end of
any mount failure. Unfortunately, that isn't true, and the change was
rolled back in D46169999.

There is still an opportunity to simplify the control flow
slightly. Replace a use of thenTry that immediately handles the error
case with a thenError, thenValue pair.

Reviewed By: kmancini

Differential Revision: D47384582

fbshipit-source-id: adccb46e2cfa4b57219fc0874f7c7fe3a7ebfbd2