Refactor code for hash types
Summary: The `hash_bytes` function is used to convert raw bytes into a Mononoke hash type (e.g. `Sha1`, `Sha256`, `Blake3`). Currently the function takes an instance of `Bytes` as argument which in some cases requires copying the content to create a new instance of `Bytes` at the caller. This diff updates the function signature to take `impl AsRef<[u8]>` instead and updates the call sites to take advantage of this change. This diff also implements `From<Sha1> for GitSha1` type

Reviewed By: markbt

Differential Revision: D47514123

fbshipit-source-id: 1ad866b438ddc4acda358caa39d267232d847563