tests: migrate test-shelve.t to modern repo
Summary:
I made the Rust clone automatically migrate to lazytext if devel.segmented-changelog-rev-compat=true. This is required to actually get legacy rev numbers counting from 0.

I also tweaked "hg share" to reload the repo. This was required for the destination repo to load the paths.default properly (and be able to initialize its edenapi).

Reviewed By: quark-zju

Differential Revision: D46832709

fbshipit-source-id: 331c7fbce5484f834c409037cdf994a10f9061a4