status: don't flush pending dirstate changes in tx
Summary: Tweak the Rust watchman status to not try flushing the dirstate if the dirstate was dirty to begin with. This avoids accidentally flushing pending dirstate changes within a Python transaction. If there are no pending changes, then it should be safe to flush treestate updates we made based on the watchman response.

Reviewed By: zzl0

Differential Revision: D46888608

fbshipit-source-id: 688a298ffb823822df2e7d342ef00d24530f759d